re-organized http_util
