+++ overload/src/overload/http_util/cluster.rs
@@ -1,18 +1,86 @@
-use crate::http_util::request::JobStatusQueryParams;
-use crate::http_util::{GenericError, GenericResponse, PATH_JOB_STATUS, PATH_STOP_JOB};
-use crate::{ErrorCode, JobStatus};
+use crate::executor::cluster::cluster_execute_request_generator;
+use crate::http_util::request::{JobStatusQueryParams, Request};
+use crate::http_util::{
+    job_id, unknown_error_resp, GenericError, GenericResponse, GenericResult, PATH_JOB_STATUS,
+    PATH_STOP_JOB,
+};
+use crate::{ErrorCode, JobStatus, Response};
 use bytes::Buf;
 use cluster_mode::{Cluster, RestClusterNode};
-use http::Request;
 use hyper::client::HttpConnector;
 use hyper::{Body, Client};
-use log::trace;
+use log::{error, trace};
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::collections::HashSet;
 use std::sync::Arc;
 
 type GenericResult<T> = Result<GenericResponse<T>, GenericError>;
+
+pub async fn handle_request_cluster(request: Request, cluster: Arc<Cluster>) -> Response {
+    let job_id = job_id(&request);
+    let buckets = request.histogram_buckets.clone();
+    if !cluster.is_active().await {
+        Response::new(job_id, JobStatus::Error(ErrorCode::InactiveCluster))
+    } else if cluster.is_primary().await {
+        let generator = request.into();
+        tokio::spawn(cluster_execute_request_generator(
+            generator,
+            job_id.clone(),
+            cluster,
+            buckets,
+        ));
+        Response::new(job_id, JobStatus::Starting)
+    } else {
+        //forward request to primary
+        let client = Client::new();
+        let job_id = request
+            .name
+            .clone()
+            .unwrap_or_else(|| "unknown_job".to_string());
+        match forward_test_request(request, cluster, client).await {
+            Ok(resp) => resp,
+            Err(err) => {
+                error!("{}", err);
+                unknown_error_resp(job_id)
+            }
+        }
+    }
+}
+
+async fn forward_test_request(
+    request: Request,
+    cluster: Arc<Cluster>,
+    client: Client<HttpConnector>,
+) -> anyhow::Result<Response> {
+    let primaries = cluster
+        .primaries()
+        .await
+        .ok_or_else(|| anyhow::anyhow!("Primary returns None"))?;
+    if let Some(primary) = primaries.iter().next() {
+        let uri = primary
+            .service_instance()
+            .uri()
+            .as_ref()
+            .ok_or_else(|| anyhow::anyhow!("Invalid ServiceInstance URI"))?;
+        trace!(
+            "forwarding request to primary: {}, {}",
+            &uri,
+            serde_json::to_string(&request).unwrap()
+        );
+        let req = hyper::Request::builder()
+            .uri(format!("{}/test", &uri))
+            .method("POST")
+            .body(Body::from(serde_json::to_string(&request)?))?;
+        let resp = client.request(req).await?;
+        let bytes = hyper::body::to_bytes(resp.into_body()).await?;
+        let resp = serde_json::from_slice::<Response>(bytes.as_ref())?;
+        return Ok(resp);
+    };
+    Ok(unknown_error_resp(
+        request.name.unwrap_or_else(|| "unknown_job".to_string()),
+    ))
+}
 
 pub async fn handle_history_all(
     params: JobStatusQueryParams,
@@ -30,7 +98,7 @@ pub async fn handle_history_all(
                 .uri(format!("{}{}?{}", &uri, PATH_JOB_STATUS, params))
                 .method("GET")
                 .body(Body::empty())?;
-            forward_test_request::<JobStatus>(request, cluster, client).await
+            forward_other_requests_to_primary::<JobStatus>(request, cluster, client).await
         } else {
             Err(GenericError::internal_500("Unknown error"))
         }
@@ -50,7 +118,7 @@ pub async fn stop(job_id: String, cluster: Arc<Cluster>) -> GenericResult<String
                 .uri(format!("{}{}/{}", &uri, PATH_STOP_JOB, &job_id))
                 .method("GET")
                 .body(Body::empty())?;
-            forward_test_request::<String>(request, cluster, client).await
+            forward_other_requests_to_primary::<String>(request, cluster, client).await
         } else {
             Err(GenericError::internal_500("Unknown error"))
         }
@@ -68,8 +136,8 @@ async fn primary_uri(
         .ok_or_else(|| GenericError::internal_500("No primary or invalid ServiceInstance URI"))
 }
 
-async fn forward_test_request<'d, T: Serialize + DeserializeOwned>(
-    req: Request<Body>,
+async fn forward_other_requests_to_primary<'d, T: Serialize + DeserializeOwned>(
+    req: http::Request<Body>,
     _cluster: Arc<Cluster>,
     client: Client<HttpConnector>,
 ) -> GenericResult<T> {
@@ -89,4 +157,8 @@ fn inactive_cluster_error() -> GenericError {
         },
         data: Default::default(),
     }
+}
+
+fn unknown_error_resp(job_id: String) -> Response {
+    Response::new(job_id, JobStatus::Error(ErrorCode::Others))
 }
